url: expose urlToHttpOptions utility
