+++ lib/_http_client.js
@@ -57,7 +57,7 @@ const { OutgoingMessage } = require('_http_outgoing');
 const Agent = require('_http_agent');
 const { Buffer } = require('buffer');
 const { defaultTriggerAsyncIdScope } = require('internal/async_hooks');
-const { URL, urlToOptions, searchParamsSymbol } = require('internal/url');
+const { URL, urlToHttpOptions, searchParamsSymbol } = require('internal/url');
 const { kOutHeaders, kNeedDrain } = require('internal/http');
 const { connResetException, codes } = require('internal/errors');
 const {
@@ -104,7 +104,7 @@ function ClientRequest(input, options, cb) {
   if (typeof input === 'string') {
     const urlStr = input;
     try {
-      input = urlToOptions(new URL(urlStr));
+      input = urlToHttpOptions(new URL(urlStr));
     } catch (err) {
       input = url.parse(urlStr);
       if (!input.hostname) {
@@ -121,7 +121,7 @@ function ClientRequest(input, options, cb) {
   } else if (input && input[searchParamsSymbol] &&
              input[searchParamsSymbol][searchParamsSymbol]) {
     // url.URL instance
-    input = urlToOptions(input);
+    input = urlToHttpOptions(input);
   } else {
     cb = options;
     options = input;
