+++ lib/internal/url.js
@@ -1295,7 +1295,7 @@ function domainToUnicode(domain) {
 // Utility function that converts a URL object into an ordinary
 // options object as expected by the http.request and https.request
 // APIs.
-function urlToOptions(url) {
+function urlToHttpOptions(url) {
   const options = {
     protocol: url.protocol,
     hostname: typeof url.hostname === 'string' &&
@@ -1494,7 +1494,7 @@ module.exports = {
   URLSearchParams,
   domainToASCII,
   domainToUnicode,
-  urlToOptions,
+  urlToHttpOptions,
   formatSymbol: kFormat,
   searchParamsSymbol: searchParams,
   encodeStr
