+++ lib/url.js
@@ -47,9 +47,10 @@ const {
   URLSearchParams,
   domainToASCII,
   domainToUnicode,
+  fileURLToPath,
   formatSymbol,
   pathToFileURL,
-  fileURLToPath
+  urlToHttpOptions,
 } = require('internal/url');
 
 // Original url.parse() API
@@ -987,5 +988,6 @@ module.exports = {
 
   // Utilities
   pathToFileURL,
-  fileURLToPath
+  fileURLToPath,
+  urlToHttpOptions,
 };
